refactor: merge some duplicated and diverged data classes
