+++ vaadin-form-layout-flow-parent/pom.xml
@@ -12,7 +12,6 @@
     <modules>
         <module>vaadin-form-layout-flow</module>
         <module>vaadin-form-layout-testbench</module>
-        <module>vaadin-form-layout-flow-demo</module>
     </modules>
     <profiles>
         <profile>
