+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/main/java/com/vaadin/flow/component/grid/it/BeanGenerator.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.grid.demo;
+package com.vaadin.flow.component.grid.it;
 
 import java.util.HashMap;
 import java.util.Map;
