+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/main/java/com/vaadin/flow/component/grid/it/GridDemo.java
@@ -75,6 +75,9 @@ import com.vaadin.flow.component.select.Select;
 import com.vaadin.flow.component.textfield.TextArea;
 import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.component.treegrid.TreeGrid;
+import com.vaadin.flow.data.bean.Address;
+import com.vaadin.flow.data.bean.Person;
+import com.vaadin.flow.data.bean.Person.MaritalStatus;
 import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.data.converter.StringToIntegerConverter;
 import com.vaadin.flow.data.provider.ListDataProvider;
@@ -94,228 +97,6 @@ import com.vaadin.flow.router.Route;
 @JsModule("@vaadin/flow-frontend/grid-demo-styles.js")
 @SuppressWarnings("squid:S1192")
 public class GridDemo extends Div {
-
-    // begin-source-example
-    // source-example-heading: Grid example model
-
-    /**
-     * Example object.
-     */
-    public static class Person implements Cloneable {
-        private int id;
-        private String firstName;
-        private String lastName;
-        private int age;
-        private Address address;
-        private String phoneNumber;
-        private MaritalStatus maritalStatus;
-        private LocalDate birthDate;
-        private boolean isSubscriber;
-        private String email;
-
-        public Person() {
-
-        }
-
-        public Person(int id, String firstName, String lastName, int age,
-                Address address, String phoneNumber) {
-            super();
-            this.id = id;
-            this.firstName = firstName;
-            this.lastName = lastName;
-            this.age = age;
-            this.address = address;
-            this.phoneNumber = phoneNumber;
-        }
-
-        public Person(int id, String firstName, String lastName, int age,
-                Address address, String phoneNumber,
-                MaritalStatus maritalStatus, LocalDate birthDate) {
-            super();
-            this.id = id;
-            this.firstName = firstName;
-            this.lastName = lastName;
-            this.age = age;
-            this.address = address;
-            this.phoneNumber = phoneNumber;
-            this.maritalStatus = maritalStatus;
-            this.birthDate = birthDate;
-        }
-
-        public int getId() {
-            return id;
-        }
-
-        public void setId(int id) {
-            this.id = id;
-        }
-
-        public String getFirstName() {
-            return firstName;
-        }
-
-        public void setFirstName(String firstName) {
-            this.firstName = firstName;
-        }
-
-        public String getLastName() {
-            return lastName;
-        }
-
-        public void setLastName(String lastName) {
-            this.lastName = lastName;
-        }
-
-        public int getAge() {
-            return age;
-        }
-
-        public void setAge(int age) {
-            this.age = age;
-        }
-
-        public String getPhoneNumber() {
-            return phoneNumber;
-        }
-
-        public void setPhoneNumber(String phoneNumber) {
-            this.phoneNumber = phoneNumber;
-        }
-
-        public Address getAddress() {
-            return address;
-        }
-
-        public void setAddress(Address address) {
-            this.address = address;
-        }
-
-        public MaritalStatus getMaritalStatus() {
-            return maritalStatus;
-        }
-
-        public void setMaritalStatus(MaritalStatus maritalStatus) {
-            this.maritalStatus = maritalStatus;
-        }
-
-        public LocalDate getBirthDate() {
-            return birthDate;
-        }
-
-        public void setBirthDate(LocalDate birthDate) {
-            this.birthDate = birthDate;
-        }
-
-        public String getImage() {
-            return "https://randomuser.me/api/portraits/men/" + getId()
-                    + ".jpg";
-        }
-
-        public boolean isSubscriber() {
-            return isSubscriber;
-        }
-
-        public void setSubscriber(boolean isSubscriber) {
-            this.isSubscriber = isSubscriber;
-        }
-
-        public String getEmail() {
-            return email;
-        }
-
-        public void setEmail(String email) {
-            this.email = email;
-        }
-
-        @Override
-        public int hashCode() {
-            return id;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj)
-                return true;
-            if (!(obj instanceof Person)) {
-                return false;
-            }
-            Person other = (Person) obj;
-            return id == other.id;
-        }
-
-        @Override
-        public String toString() {
-            return String.format("%s, %s", firstName, lastName);
-        }
-
-        @Override
-        public Person clone() { // NOSONAR
-            try {
-                return (Person) super.clone();
-            } catch (CloneNotSupportedException e) {
-                throw new RuntimeException(
-                        "The Person object could not be cloned.", e);
-            }
-        }
-    }
-
-    public static class Address {
-        private String street;
-        private int number;
-        private String postalCode;
-        private String city;
-
-        public Address() {
-
-        }
-
-        public Address(String postalCode, String city) {
-            this.postalCode = postalCode;
-            this.city = city;
-        }
-
-        public String getStreet() {
-            return street;
-        }
-
-        public void setStreet(String street) {
-            this.street = street;
-        }
-
-        public int getNumber() {
-            return number;
-        }
-
-        public void setNumber(int number) {
-            this.number = number;
-        }
-
-        public String getPostalCode() {
-            return postalCode;
-        }
-
-        public void setPostalCode(String postalCode) {
-            this.postalCode = postalCode;
-        }
-
-        public String getCity() {
-            return city;
-        }
-
-        public void setCity(String city) {
-            this.city = city;
-        }
-
-        @Override
-        public String toString() {
-            return String.format("%s %s", postalCode, city);
-        }
-
-    }
-
-    public enum MaritalStatus {
-        MARRIED, SINGLE;
-    }
 
     // end-source-example
 
@@ -612,7 +393,7 @@ public class GridDemo extends Div {
 
     }
 
-    protected GridDemo() {
+    public GridDemo() {
         createBasicUsage();// Basic Grid
         addVariantFeature();
         createGridWithLazyLoading(); // Lazy Loading
@@ -1173,7 +954,7 @@ public class GridDemo extends Div {
         TextField postalCodeField = new TextField();
         postalCodeField.addValueChangeListener(
                 event -> dataView.addFilter(person -> StringUtils
-                        .containsIgnoreCase(person.getAddress().getPostalCode(),
+                        .containsIgnoreCase(person.getAddress().getPostalCode().toString(),
                                 postalCodeField.getValue())));
 
         postalCodeField.setValueChangeMode(ValueChangeMode.EAGER);
