+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/main/java/com/vaadin/flow/component/grid/it/entity/Customer.java
@@ -1,4 +1,4 @@
-package com.vaadin.flow.component.grid.demo.entity;
+package com.vaadin.flow.component.grid.it.entity;
 
 public class Customer {
     private int id;
