+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/main/java/com/vaadin/flow/component/grid/it/entity/Task.java
@@ -1,4 +1,4 @@
-package com.vaadin.flow.component.grid.demo.entity;
+package com.vaadin.flow.component.grid.it.entity;
 
 import java.time.LocalDate;
