+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/main/java/com/vaadin/flow/data/bean/Person.java
@@ -2,6 +2,7 @@ package com.vaadin.flow.data.bean;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.util.Date;
 
 public class Person implements Serializable, Cloneable {
@@ -12,6 +13,8 @@ public class Person implements Serializable, Cloneable {
     private int age;
     private Gender gender;
     private Address address;
+    private String phoneNumber;
+    private MaritalStatus maritalStatus;
     private boolean deceased;
     private Date birthDate;
     private boolean isSubscriber;
@@ -23,6 +26,31 @@ public class Person implements Serializable, Cloneable {
 
     public Person() {
 
+    }
+
+    public Person(int id, String firstName, String lastName, int age,
+            Address address, String phoneNumber) {
+        super();
+        this.id = id;
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.age = age;
+        this.address = address;
+        this.phoneNumber = phoneNumber;
+    }
+
+    public Person(int id, String firstName, String lastName, int age,
+            Address address, String phoneNumber,
+            MaritalStatus maritalStatus, Date birthDate) {
+        super();
+        this.id = id;
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.age = age;
+        this.address = address;
+        this.phoneNumber = phoneNumber;
+        this.maritalStatus = maritalStatus;
+        this.birthDate = birthDate;
     }
 
     public int getId() {
@@ -155,6 +183,27 @@ public class Person implements Serializable, Cloneable {
         this.birthDate = birthDate;
     }
 
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public MaritalStatus getMaritalStatus() {
+        return maritalStatus;
+    }
+
+    public void setMaritalStatus(MaritalStatus maritalStatus) {
+        this.maritalStatus = maritalStatus;
+    }
+
+    public String getImage() {
+        return "https://randomuser.me/api/portraits/men/" + getId()
+                + ".jpg";
+    }
+
     @Override
     public Person clone() {
         try {
@@ -174,5 +223,9 @@ public class Person implements Serializable, Cloneable {
         return new Person("Maya", "Dinkelstein", "maya@foo.bar", 18,
                 Gender.FEMALE, new Address("Red street", 12, "Amsterdam",
                         Country.NETHERLANDS));
+    }
+
+    public enum MaritalStatus {
+        MARRIED, SINGLE;
     }
 }
