+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/main/resources/META-INF/resources/frontend/grid-demo-styles.js
